+++ src/bridge_lcd.h
@@ -37,7 +37,21 @@
 #define GFXFF                   1
 #define TILTS_PER_PAGE          5 // The actual number is one fewer than this - the first row is used for headers
 
+#elif defined(LCD_TFT_M5STICKC)
+
+#include <M5StickC.h>
+
+#define TFT_M5STICKC_LINE_CLEARANCE 0
+#define GFXFF                   2
+#define TILTS_PER_PAGE          5 // The actual number is one fewer than this - the first row is used for headers
+
 #endif // LCD_SSD1306
+
+#ifdef TILTS_PER_PAGE
+#define HAVE_LCD                1
+#else
+#define HAVE_LCD                0
+#endif
 
 #define SCREEN_TILT             0
 #define SCREEN_LOGO             1
@@ -68,6 +82,7 @@ public:
     void clear();
 
 private:
+#if HAVE_LCD
     uint8_t display_next();
     void display_tilt_screen(uint8_t screen_number);
     void print_tilt_to_line(tiltHydrometer *tilt, uint8_t line);
@@ -76,9 +91,7 @@ private:
 
 #ifdef LCD_SSD1306
     SSD1306Wire *oled_display;
-#elif defined(LCD_TFT)
-    TFT_eSPI *tft;
-#elif defined(LCD_TFT_ESPI)
+#elif defined(LCD_TFT) || defined(LCD_TFT_ESPI) || defined(LCD_TFT_M5STICKC)
     TFT_eSPI *tft;
 #endif // LCD_SSD1306
 
@@ -88,6 +101,7 @@ private:
     unsigned long next_screen_at;
 
     bool touchLatch = false;    // Ensure we only trigger a touch once
+#endif // HAVE_LCD
 };
 
 void screenFlip();