+++ docs/source/changelog.rst
@@ -10,6 +10,7 @@ v1.0.3 - Unreleased
 - Added 24 hour reboot timer to improve overall stability
 - Shorten logo display time by 60%
 - Added reboot if reconnecting to WiFi fails
+- Added Grainfather as a push target (Thanks @teanooki!)
 
 
 v1.0.2 - Feb 19, 2021 - Calibration Bugfixes